cs changes

+++ Extensions/ActionManagerExtensions.cs
@@ -8,7 +8,7 @@ public static unsafe class ActionManagerExtensions {
         public static bool IsCasting(ActionType actionType, uint actionId) {
             if (Svc.Objects.LocalPlayer is not { } player) return false;
             var info = player.Address.As<Character>()->GetCastInfo();
-            return info is not null && info->IsCasting && info->ActionType == actionType && info->ActionId == actionId;
+            return info is not null && info->IsCasting && info->ActionType == (byte)actionType && info->ActionId == actionId;
         }
 
         public static bool UseAction(ActionType actionType, uint actionId) {
